requirements.txt
Cerberus==1.3.7
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
paho-mqtt==2.1.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
urllib3==2.4.0
